requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
pandas
optuna
mlflow
gradio
streamlit
requests
openai
scikit-learn
faiss-cpu
transformers
langchain
sentence-transformers
pdfplumber
python-multipart
shap
prometheus_fastapi_instrumentator